requirements.txt
pandas
pillow
reportlab
openpyxl
pyinstaller
